requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
python-multipart>=0.0.20,<1.0.0
PyMuPDF>=1.26.0,<2.0.0
python-docx>=1.1.2,<2.0.0
